runtime: split get_first_tag_in_range QA out of tag_utils
it's separate.
