+++ gnuradio-runtime/python/gnuradio/gr/CMakeLists.txt
@@ -30,8 +30,14 @@ if(ENABLE_TESTING)
                             ${PROJECT_BINARY_DIR}/gnuradio-runtime/python)
     include(GrTest)
     file(GLOB py_qa_test_files "qa_*.py")
-    set(py_qa_test_files qa_flowgraph.py qa_prefs.py qa_python_logging.py qa_random.py
-                         qa_sys_paths.py qa_tag_utils.py)
+    set(py_qa_test_files    qa_flowgraph.py
+                            qa_prefs.py
+                            qa_python_logging.py
+                            qa_random.py
+                            qa_sys_paths.py
+                            qa_tag_utils.py
+                            qa_get_first_tag_in_range.py
+    )
     # This is a check for whether gr-blocks is enabled
     if(ENABLE_DEFAULT OR ENABLE_GR_BLOCKS)
         list(APPEND py_qa_test_files qa_hier_block2.py qa_uncaught_exception.py)
